move Client API code into Network.Consul.Client.*
For example, the code to interact with the Session API has been moved to the `Network.Consul.Client.Session` module.

This moves code for the following APIs:
* Agent
* Catalog
* Health
* KV Store
* Session

as well as the additional code in:
* Client Init
* Sequencer
* runService helper

This change doesn't break the API, all modules and interfaces are still exported/exposed via the `Network.Consul` module.
